renamed Navbar Topbar for specificity and not to conflict with reactstrap Navbar
