+++ frontend/src/components/Content/Content.js
@@ -1,11 +1,11 @@
 import React from "react"
 import classNames from "classnames"
 import { Container } from "reactstrap"
-import NavBar from "./Navbar"
+import Topbar from "./Topbar"
 
 export default ({ isOpen, toggle, children }) => (
   <Container fluid className={classNames("content", { "is-open": isOpen })}>
-    <NavBar toggle={toggle} />
+    <Topbar toggle={toggle} />
     {children}
   </Container>
 )
